pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["dbp"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
coverage = "^6.4.4"
Markdown = "^3.4.1"
django-filter = "^22.1"
drf-yasg = "^1.21.3"

[tool.poetry.dev-dependencies]
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
freezegun = "^1.2.2"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::django.utils.deprecation.RemovedInDjango50Warning",
]